Move to monorepo
Yolo-ing this in without review. 

In developing #9, I hosted my NextJS project in `src/examples`. This "project Russian doll" approach was painful, and that pain would only have been magnified if I'd merged that PR as-is and subjected other contributors to it. So instead, I'm switching us to use a monorepo. Then I'll be able to add the NextJS project as a sibling / consumer of `ai-jsx`, which should allow the NextJS project to be less coupled.

I used [turborepo](https://turbo.build/repo) as the monorepo runner, because I suspect it'll provide better support than raw yarn workspaces. Turborepo has a lot of perf optimizations which we don't need right now, but by being on Turborepo from the start, we'll get that for free as the project scales.

This PR also adds a GH action to test PRs.

This PR may cause merge conflicts with the many in-flight PRs; I think we just need to bite that bullet eventually though.