Revamp project structure (#22)
In this PR, we split `examples` into a separate monorepo package. This forces us to set up the `ai-jsx` package such that it exports something that can be used by a consuming module. It also prompted me to declare a delineation between `core` and `batteries`. (I'm open to naming suggestions for the latter.)