Redo NextJS Demo (#23)
This is like #9 but targeting our new project layout.

See the note in the root-level README for how to run the demo.

There are a bunch of hacks / type issue which are documented in `packages/nextjs-demo/README.md`. I think we can work around them eventually but it's not worthwhile to do so now.