Fix build system (again) (#41)
![image](https://github.com/fixie-ai/ai-jsx/assets/829827/b74a093e-e8b8-4d69-b8e0-66aa8485eb82)

> The same thing we do every day, Pinky. Fiddle with ESM + CJS build settings.

The thing that makes this tricky is we have to navigate the combinatorial explosion of:

* JS vs. TS
* `"type": "module"` caller vs. `.mjs` caller vs. CJS caller
* Within TS: ESM output vs. CJS output

I made https://github.com/NickHeiner/esm-cjs-test to exercise those cases. I think we finally have something that'll work on a reasonable set of cases.

With the current configuration, `import`s of LangChain caused this TS error:

```
TS1479: The current file is a CommonJS module whose imports will produce 'require' calls; however, the referenced file is an ECMAScript module and cannot be imported with 'require'. Consider writing a dynamic 'import("langchain/text_splitter")' call instead.
  To convert this file to an ECMAScript module, add the field `"type": "module"` to '/Users/nth/code/ai-jsx/packages/ai-jsx/package.json'.

5 import { TokenTextSplitter } from 'langchain/text_splitter';
```

However, if I `ts-expect-error` it, the types still flow through in the program. And both the generated ESM and CJS files work. So I think this is just a TS bug.

I believe this bug will apply to our callers as well.