Stream more in create-react-app (#70)
This expands the UI bridge affordance, then updates the Docs Chat example to use it.

As a result, we now have streaming results in Docs Chat. Docs Chat is also vastly simplified.

This will cause some merge conflicts with #69 but I think we mostly changed different parts so it shouldn't be too bad to resolve.

As a follow-up, it would be good to do this for the "Basic Chat" example as well.