Add support for append-only text streaming (#159)
This adds underlying support for rendering to an append-only stream.

Components that will yield in an append-only manner can do so by
yielding a special symbol, after which point all yielded values will be
appended to, rather than replace, previously yielded values. (This
indicates to rendering that it is safe to present intermediate content
when append-only rendering is requested.)

This adds an API endpoint that demonstrates responding with a text
stream that is compatible with Vercel's AI SDK but does not add any
client-side demos.