Add chat demo using server actions (#210)
This adds a chat demo that uses server actions to the
nextjs-experimental demo.

In `ai-jsx/experimental/next`, this:
- Fixes an issue with streaming where suffixes were not correctly being
computed
- Adds an `onComplete` prop to `AI.JSX` so that consumers can access the
final result (either as a `string` or as a `ReactNode`)
- Adds an experimental `stream` prop that can be used to work around the
low streaming framerate caused by `<Suspense>` throttling (at the
expense of latency)
- Adds a set of components under `ai-jsx/experimental/next/chat`
(`<Chat>`, `<Conversation>`, `<Message>`) that try to put the chat data
flow somewhat on rails. Notably, a chat UX can be constructed entirely
with server components and server actions -- AI.JSX includes the client
components that do all the data plumbing.