Allow UseTools to augment a broader conversation (#215)
This allows `<UseTools>` (specifically when using a `<ChatCompletion>`
that supports function calls) to drive the function-calling loop as part
of existing conversation. That is, it now accepts a full conversation
and invokes functions/appends to the conversation as the model requests.

It also adds a `showSteps` parameter which allows for streaming by also
streaming the intermediate steps. Without this, we don't know whether an
assistant message is terminal until it's fully rendered. (I suspect
we'll iterate on the presentation of these intermediate steps shortly.)