Change ChatCompletion to emit AssistantMessage (#218)
This changes the various `ChatCompletion` implementations to emit
`<AssistantMessage>`s rather than loose text. This allows callers to
easily separate `<AssistantMessage>` and `<FunctionCall>` outputs.

Preserving the append-only nature of the stream requires non-trivial
gymnastics to split the iterator consumption across multiple components,
but this technique also allows for the removal of the
`experimental_streamFunctionCallOnly` flag.