First stab at JS-based CLI for Fixie. (#247)
The idea here is to replace the Python-based "fixie" CLI with an
npm-based "fixie" CLI, which will allow one to create and manage
Sidekicks and Corpora. This way an AI.JSX developer need not have a
Python environment set up. It also provides a JS-based FixieClient class
for interacting with the Fixie platform APIs programmatically.

This version only has the Corpus-related commands, we will need to port
over the logic from the Python-based "fixie deploy" command for AI.JSX
Sidekick deployment to JS as well.