Rewrite use-fixie hook (#363)
Previously, all state lived in the `useFixie` hook. This was bad,
because if the calling component unmounted, all the state would be lost.
This led to flickering and an inability to reliably cache.

To solve this, we do the same thing libraries like react-query or Apollo
do: have state live outside the hook itself. This allows the hook to
simply be a shim to the React lifecycle.

This also provides an opening for other frameworks to integrate.

I'm completely confident this is an improvement in correctness over what
we have today. There may be some minor issues, but there are also some
backend issues right now, so it's a little hard to tell. I'm going to
land this and we can keep testing with it.