Add support for inferring custom runtime parameters (#366)
This adds CLI support for inferring/setting the schema for custom
runtime parameters. We use `typescript-json-schema` to infer the schema
and set it in GraphQL. (This change does not add CLI support for setting
default runtime parameters, but adds it to the underlying GraphQL
mutation.)

The bulk of the change consists of refactoring `fixie serve` so that
_either_ a tunnel URL change or parameter schema change triggers a new
revision.