TTS: orphan cancelled requests, reconnect websocket when needed (#414)
This avoids two issues:
- first, we would still send and download requests even after calling
`tts.stop()`, which would lead to errors when calling
`outputManager.appendBuffer` for a stream that no longer existed. This
now cancels all requests on stop().
- second, if a websocket closed due to idling, the TTS object would be
dead at that point. Now, we force a reconnect if the WS is closed (which
will have extra latency, which can be avoided using the new `warmup()`
method if we know a generation is imminent).