Improve prompt/completion logging (#506)
Currently the prompt/completion log events include debug-serialized JSX.
But this is a) difficult to extract data from and b) unnecessarily
generic because we're already working with a very finite set of
conversation components.

This changes the structure to include the final text and props
(sans-`children`) instead. Additionally, we include a `getElement()`
function so that consumers have the option to access the element itself,
without it being JSON-serialized by default.